controlclient: extract State and Status stuff into its own file.
No changes other than moving stuff around.
